Feature/sequence tpnr

+++ src/main/kotlin/no/nav/samordning/hendelser/feed/Feed.kt
@@ -2,4 +2,4 @@ package no.nav.samordning.hendelser.feed
 
 import no.nav.samordning.hendelser.hendelse.Hendelse
 
-class Feed(var hendelser: List<Hendelse>?, var sisteSekvensnummer: Int?, var sisteLesteSekvensnummer: Int?, var nextUrl: String?)
+class Feed(var hendelser: List<Hendelse>?, var sisteSekvensnummer: Long?, var sisteLesteSekvensnummer: Long?, var nextUrl: String?)
